refactor: [#40] Refactored project structure
