+++ src/context/types.ts
@@ -1,4 +1,4 @@
-import { IMovie, IShow } from '@/types/shared';
+import { IMovie, IShow } from '@/pages/Home/types';
 
 export type Context = {
   children: React.ReactNode;
@@ -12,3 +12,9 @@ export interface IAppContext {
   contentType: string;
   search: string;
 }
+
+export type ReducerAction =
+  | { type: 'SET_LOADING'; loading: boolean }
+  | { type: 'SET_CONTENT_TYPE' | 'SET_ACTIVE_QUERY_TYPE' | 'SET_SEARCH'; [key: string]: string }
+  | { type: 'SET_SHOWS'; shows: IShow[] }
+  | { type: 'SET_MOVIES'; movies: IMovie[] };
