+++ src/helpers/getData.ts
@@ -1,4 +1,4 @@
-import handleResponse from './handleResponse';
+import { handleResponse } from './';
 
 const getData = async <T>(url: string): Promise<T> => {
   const response = await fetch(url);
