+++ src/pages/Home/Home.tsx
@@ -5,9 +5,10 @@ import Loading from '@/components/Loading';
 import Navbar from '@/components/Navbar';
 import { CONTENT_TYPE } from '@/constants/constantValues';
 import { MoviesShowsContext } from '@/context';
-import { IShow, IMovie, IItem } from '@/types/shared';
+import { IItem } from '@/types';
 
 import './Home.css';
+import { IMovie, IShow } from './types';
 
 export default function Home() {
   const { movies, shows, contentType, loading } = useContext(MoviesShowsContext);
