+++ src/pages/MovieOrShow/MovieOrShow.tsx
@@ -7,8 +7,8 @@ import Loading from '@/components/Loading';
 import Vote from '@/components/Vote';
 import { IMAGE_PATH, DEFAULT_IMAGE, VIDEO_PATH } from '@/constants/constantValues';
 import { MoviesShowsContext } from '@/context';
-import calculateAverageVote from '@/helpers/calculateAverageVote';
-import { ContentType, IItem } from '@/types/shared';
+import { calculateAverageVote } from '@/helpers';
+import { ContentType, IItem } from '@/types';
 
 import './MovieOrShow.css';
