+++ src/pages/Registration/types.ts
@@ -1,0 +1,11 @@
+export interface IUser {
+  firstName?: string;
+  lastName?: string;
+  email: string;
+  password: string;
+}
+
+export interface IRegistrationResponse {
+  code: number;
+  message: string;
+}
